+++ Solution_finder/utility.h
@@ -21,7 +21,7 @@ static double RandNum() { return static_cast <double> (rand()) / static_cast <do
 static double GetTotalDif(std::vector<double> input, std::vector<double> output)
 {
 	double total = 0;
-	for (size_t i = 0; i < input.size(); i++)
+	for (unsigned int i = 0; i < input.size(); i++)
 		total += abs(input[i] - output[i]);
 	return total;
 }
